refactor(test) Extract shared MockImporter to conftest
why: MockImporter and CapturingMockImporter are test helpers that can be shared across CLI test files.
what:
- Move MockImporter and CapturingMockImporter to tests/cli/conftest.py
- Update test_import_repos.py to import from conftest
